added working PHP code for 3 of the 4 boxes
It's pretty slow on the original B, but it does work.  I also added
code to run gravity.sh when you click update, but I'm not quite sure
how to use the font awesome icons to show a spinner while it is
working...

All of the PHP snippets are based off of the chroniton.sh script that
is used to display data on a 2.8" touchscreen.