Add update notifications
Using the Github API, it checks the latest release against
the current version. If it needs updating, it shows an
update alert in the top right dropdown-menu and opens it
automatically. Also this commit stops the dropdown-menu
from closing if you click inside of it. It will only close
if you click the toggle button or outside of it.