Updated Fullscreen mode to be font/svg agnostic
For fullscreen icons to work you need place them within an anchor with the class="fullscreen". Then for each icon add the class="fullscreent-icon" this can be just an icon or a div that surrounds some other item, or you can add it to a svg tag. The last past for the toggle to work is to add the class="lte-show" the the same item that has the fullscreen-icon. I tested with FontAwesom icon and with heroicon svg's.